# Faster primitive arrays encoding into row format