@alamb 2024-06-11
I took the liberty of merging this PR up from main to pick up the fixes for the MSRV CI chekc