@alamb 2024-06-12
I am running benchmarks now on a server