@korowa 2024-06-13
@alamb @tustvold sorry for the late response.

Actually benchmark results are a bit unexpected (since encoding performance is the same for nullable and non-nullable columns), so I'd like to verify them, and also work on @tustvold 's comments.

Thank you.