@alamb 2024-06-13
To make running benchmarks easier, perhaps we could pull them out into their own PR that we could merge into `master` independently to this PR