@korowa 2024-06-18
@alamb @tustvold I suppose this version to be ready for merging -- could you, please, take a look when you have time?

And regarding benchmark results -- all the changes affect only `convert_columns` benchmarks, `convert_rows` ones should not be affected by this PR.