@alamb 2024-06-19
> @alamb @tustvold I suppose this version to be ready for merging -- could you, please, take a look when you have time?

Thanks @korowa  -- I am running the benchmarks again and I'll give this another look