@alamb 2024-06-19
🤔  I reran the benchmarks and it actually looks to me like this branch is *slower* than main (and looking at the results in https://github.com/apache/arrow-rs/pull/5858#pullrequestreview-2113650429 I think I misread them and they show this branch is slower too)

Here are the results:

```
group                                                                                                                         encode-skip-iter                       master
-----                                                                                                                         ----------------                       ------
convert_rows 4096 bool(0, 0.5)                                                                                                1.15     27.6±0.04µs        ? ?/sec    1.00     23.9±0.03µs        ? ?/sec
convert_rows 4096 bool(0.3, 0.5)                                                                                              1.15     27.7±0.04µs        ? ?/sec    1.00     24.0±0.29µs        ? ?/sec
convert_rows 4096 i64(0)                                                                                                      1.18     47.5±0.06µs        ? ?/sec    1.00     40.4±0.03µs        ? ?/sec
convert_rows 4096 i64(0.3)                                                                                                    1.18     47.7±0.18µs        ? ?/sec    1.00     40.6±0.46µs        ? ?/sec
convert_rows 4096 u64(0)                                                                                                      1.19     47.3±0.05µs        ? ?/sec    1.00     39.6±0.09µs        ? ?/sec
convert_rows 4096 u64(0.3)                                                                                                    1.19     47.3±0.06µs        ? ?/sec    1.00     39.7±0.13µs        ? ?/sec
```

Here is the script I used:
```shell
git checkout encode-skip-iter
cargo bench --all-features --bench row_format -- --save-baseline encode-skip-iter

git reset `git merge-base HEAD apache/master`
git restore arrow-row/src/lib.rs
git reset git restore arrow-row/src/fixed.rs
git status

cargo bench --all-features --bench row_format -- --save-baseline master

critcmp master encode-skip-iter
```

I am rerunning to be sure